ref #21 added write module to antiweb_lib package
